requirements.txt
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth==2.25.2
cryptography==41.0.7
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
supervisor==4.2.5